take a number of dimensions as argument
This PR is meant as a the start of a conversation on PresetInterpolatorServer's usage. The change I am proposing above will probably break existing code, and therefore should not be merged. Think of it as a suggestion that can be improved upon :)

I propose that PrIntSvr take a number of dimensions as argument when created. Ideally, I think that the number of dimensions should be the only argument. This seems to work when omitting the model argument:
```supercollider
PresetInterpolatorServer(dimensions: 3);
```